Added database & connection. Datalaag werkt volledig.

+++ src/data/migrations/2021112260_addPercentageToBeers.js
@@ -3,7 +3,7 @@ const { tables } = require('..');
 module.exports = {
   up: async (knex) => {
     await knex.schema.alterTable(tables.beer, (table) => {
-      table.double('percentage')
+      table.integer('percentage')
         .nullable();
     });
   },
